Adding drive to robots
