+++ Assets/Packages.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 70e5e67b2317e5f08acf8e93a732912f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
